backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.19.1
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.7.0
fsspec==2025.9.0
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.2.1
numpy==1.24.3
openai==1.107.0
packaging==25.0
pillow==10.2.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.31.0
safetensors==0.6.2
sniffio==1.3.1
sympy==1.14.0
timm==0.9.12
torch==2.1.0
torchvision==0.16.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
